split recursive logic tests into a separate file
