+++ packages/forms/experimental/test/node/field_node.spec.ts
@@ -11,12 +11,9 @@ import {TestBed} from '@angular/core/testing';
 import {
   apply,
   applyEach,
-  applyWhen,
   disabled,
   FieldPath,
   form,
-  MIN,
-  min,
   readonly,
   required,
   REQUIRED,
@@ -28,11 +25,6 @@ import {
 } from '../../public_api';
 import {ValidationError, WithField} from '../../src/api/validation_errors';
 import {SchemaImpl} from '../../src/schema/schema';
-
-interface TreeData {
-  level: number;
-  next: TreeData;
-}
 
 describe('FieldNode', () => {
   it('can get a child of a key that exists', () => {
@@ -881,113 +873,6 @@ describe('FieldNode', () => {
       );
 
       expect(() => f().disabled()).toThrowError('Path is not part of this field tree.');
-    });
-
-    it('should support recursive logic', () => {
-      const s = schema<TreeData>((p) => {
-        disabled(p.level, ({valueOf}) => {
-          return valueOf(p.level) % 2 === 0;
-        });
-        apply(p.next, s);
-      });
-      const f = form<TreeData>(
-        signal({level: 0, next: {level: 1, next: {level: 2, next: {level: 3, next: null!}}}}),
-        s,
-        {injector: TestBed.inject(Injector)},
-      );
-      expect(f.level().disabled()).toBe(true);
-      expect(f.next.level().disabled()).toBe(false);
-      expect(f.next.next.level().disabled()).toBe(true);
-      expect(f.next.next.next.level().disabled()).toBe(false);
-    });
-
-    it('should support co-recursive logic', () => {
-      const s1: Schema<TreeData> = schema((p) => {
-        disabled(p.level, ({valueOf}) => valueOf(p.level) % 2 === 0);
-        apply(p.next, s2);
-      });
-      const s2: Schema<TreeData> = schema((p) => {
-        disabled(p.level, ({valueOf}) => valueOf(p.level) % 2 === 0);
-        apply(p.next, s1);
-      });
-      const f = form<TreeData>(
-        signal({
-          level: 0,
-          next: {level: 1, next: {level: 2, next: {level: 3, next: null!}}},
-        }),
-        s1,
-        {injector: TestBed.inject(Injector)},
-      );
-      expect(f.level().disabled()).toBe(true);
-      expect(f.next.level().disabled()).toBe(false);
-      expect(f.next.next.level().disabled()).toBe(true);
-      expect(f.next.next.next.level().disabled()).toBe(false);
-    });
-
-    it('should support recursive logic terminated by a when condition', () => {
-      const s: Schema<TreeData> = schema((p) => {
-        min(p.level, ({valueOf}) => valueOf(p.level));
-        applyWhen(p.next, (ctx) => ctx.valueOf(p.level) !== 2, s);
-      });
-      const f = form<TreeData>(
-        signal({
-          level: 0,
-          next: {level: 1, next: {level: 2, next: {level: 3, next: {level: 4, next: null!}}}},
-        }),
-        s,
-        {injector: TestBed.inject(Injector)},
-      );
-      expect(f.level().property(MIN)()).toBe(0);
-      expect(f.next.level().property(MIN)()).toBe(1);
-      expect(f.next.next.level().property(MIN)()).toBe(2);
-      expect(f.next.next.next.level().property(MIN)()).toBe(undefined);
-      expect(f.next.next.next.next.level().property(MIN)()).toBe(undefined);
-    });
-
-    it('should support recursive logic with arrays', () => {
-      interface Dom {
-        tag: string;
-        children: Dom[];
-      }
-
-      const domSchema = schema<Dom>((p) => {
-        applyEach(p.children, domSchema);
-        applyWhen(
-          p.children,
-          ({valueOf}) => valueOf(p.tag) === 'table',
-          (children) => {
-            applyEach(children, (c) => {
-              validate(c.tag, ({value}) =>
-                value() !== 'tr' ? ValidationError.custom({kind: 'invalid-child'}) : undefined,
-              );
-            });
-          },
-        );
-        applyWhen(
-          p.children,
-          ({valueOf}) => valueOf(p.tag) === 'tr',
-          (children) => {
-            applyEach(children, (c) => {
-              validate(c.tag, ({value}) =>
-                value() !== 'td' ? ValidationError.custom({kind: 'invalid-child'}) : undefined,
-              );
-            });
-          },
-        );
-      });
-
-      const data = signal<Dom>({tag: 'div', children: [{tag: 'span', children: []}]});
-      const f = form(data, domSchema, {injector: TestBed.inject(Injector)});
-      expect(f().valid()).toBe(true);
-
-      data.set({tag: 'table', children: [{tag: 'span', children: []}]});
-      expect(f().valid()).toBe(false);
-
-      data.set({tag: 'table', children: [{tag: 'tr', children: [{tag: 'span', children: []}]}]});
-      expect(f().valid()).toBe(false);
-
-      data.set({tag: 'table', children: [{tag: 'tr', children: [{tag: 'td', children: []}]}]});
-      expect(f().valid()).toBe(true);
     });
   });
